Fix wrong data types (int instead of float in db)

+++ mags-server/src/main/java/com/jacobsonmt/mags/server/model/result/MaGSResult.java
@@ -48,7 +48,7 @@ public class MaGSResult {
     private Double disorder;
 
     private Double compositionD;
-    private Integer compositionE;
+    private Double compositionE;
 
     private Double compositionL;
     private Double compositionG;
